move math into its own class
