+++ client/app/styles/queue/_evaluate_decision.scss
@@ -4,6 +4,10 @@
   .case-type {
     width: 15%;
   }
+
+  .attorney-assigned {
+    width: 30%;
+  }
 }
 
 .grayLineStyling {
